requirements.txt
sentence-transformers
anndata>=0.6.4
fbpca>=1.0
umap-learn>=0.3.10
matplotlib>=2.0.2
numpy>=1.16.4
scipy>=1.3.1
scikit-learn>=0.21.3
scanorama>=1.7.1
torch==1.7.1
torchvision==0.8.2
torchaudio==0.7.2
torchvision==0.8.2
pandas
tqdm
networkx
gensim
nltk